splitting Device and Device Manager
